add mongo driver and upgrade go version to 1.18
