+++ .gitignore
@@ -2,3 +2,4 @@ coverage.out
 vendor/*
 !/vendor/vendor.json
 /gorm/test.db
+.idea
